Extract the new context menu changes out of the ADS
