+++ app/src/main/res/layout/popup_long_press_image_link_menu.xml
@@ -46,7 +46,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:leadingIcon="@drawable/ic_downloads_24"
-        app:popupLeadingIconSize="24dp"
         app:primaryText="@string/downloadImage" />
 
     <com.duckduckgo.common.ui.view.PopupMenuItemView
@@ -54,7 +53,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:leadingIcon="@drawable/ic_open_in_24"
-        app:popupLeadingIconSize="24dp"
         app:primaryText="@string/openImageInNewTab" />
 
     <com.duckduckgo.common.ui.view.divider.HorizontalDivider
@@ -67,7 +65,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:leadingIcon="@drawable/ic_add_24"
-        app:popupLeadingIconSize="24dp"
         app:primaryText="@string/openInNewTab" />
 
     <com.duckduckgo.common.ui.view.PopupMenuItemView
@@ -75,7 +72,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:leadingIcon="@drawable/ic_open_in_24"
-        app:popupLeadingIconSize="24dp"
         app:primaryText="@string/openInBackgroundTab" />
 
     <com.duckduckgo.common.ui.view.divider.HorizontalDivider
@@ -89,7 +85,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:leadingIcon="@drawable/ic_fire_tabs_24"
-        app:popupLeadingIconSize="24dp"
         app:primaryText="@string/openInFireTab" />
 
     <com.duckduckgo.common.ui.view.divider.HorizontalDivider
@@ -103,7 +98,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:leadingIcon="@drawable/ic_copy_24"
-        app:popupLeadingIconSize="24dp"
         app:primaryText="@string/copyUrl" />
 
     <com.duckduckgo.common.ui.view.PopupMenuItemView
@@ -111,6 +105,5 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:leadingIcon="@drawable/ic_share_android_24"
-        app:popupLeadingIconSize="24dp"
         app:primaryText="@string/shareLink" />
 </LinearLayout>
