+++ app/src/main/res/values/dimens.xml
@@ -31,4 +31,5 @@
     <dimen name="browserModeToggleSegmentWidth">56dp</dimen>
     <dimen name="browserModeToggleSegmentHeight">40dp</dimen>
     <dimen name="longPressMenuWidth">240dp</dimen>
+    <dimen name="longPressMenuLeadingIconSize">24dp</dimen>
 </resources>
